change _id from uint8_t to int to get rid of compiler warnings

+++ stepper_motor_controller/tc74.h
@@ -72,7 +72,7 @@ public:
     }
 
 private:
-    uint8_t _id;
+    int _id;
 };
 
 #endif /* TC74_H_ */
